quant_soxl_bot/requirements.txt
backtrader
pandas
numpy
matplotlib
alpaca-py
python-dotenv
